node leaf work

+++ src/leaf/modes/PaletteMode.cpp
@@ -1,14 +1,14 @@
 #include "PaletteMode.h"
 
-CPaletteMode::CPaletteMode(const uint16_t numLeds, const TProgmemRGBPalette16& palette, const uint16_t increment)
+CPaletteMode::CPaletteMode(const uint16_t numLeds, const TProgmemRGBPalette16& palette, const float increment)
 : CBaseMode(numLeds), increment(increment), palette(palette), blendType(LINEARBLEND), delay(15) {
 }
 
-CPaletteMode::CPaletteMode(const uint16_t numLeds, const TProgmemRGBPalette16& palette, const uint16_t increment, const TBlendType blendType)
+CPaletteMode::CPaletteMode(const uint16_t numLeds, const TProgmemRGBPalette16& palette, const float increment, const TBlendType blendType)
 : CBaseMode(numLeds), increment(increment), palette(palette), blendType(blendType), delay(15) {
 }
 
-CPaletteMode::CPaletteMode(const uint16_t numLeds, const TProgmemRGBPalette16& palette, const uint16_t increment, const TBlendType blendType, const unsigned long delay)
+CPaletteMode::CPaletteMode(const uint16_t numLeds, const TProgmemRGBPalette16& palette, const float increment, const TBlendType blendType, const unsigned long delay)
 : CBaseMode(numLeds), increment(increment), palette(palette), blendType(blendType), delay(delay) {
 }
 
@@ -20,9 +20,9 @@ void CPaletteMode::loop() {
 }
 
 void CPaletteMode::draw(CRGB *leds) {
-    uint8_t ci = startIndex;
+    float ci = startIndex;
     for( int i = 0; i < numLeds; i++) {
-        leds[i] = ColorFromPalette( palette, ci, 255 * configuration.ledBrightness, blendType);
+        leds[i] = ColorFromPalette( palette, (uint8_t)ci, 255 * configuration.ledBrightness, blendType);
         ci+=increment;
     }
 }
